+++ app/src/main/java/org/simple/clinic/user/LoggedInUserHttpInterceptor.kt
@@ -1,42 +1,33 @@
 package org.simple.clinic.user
 
-import dagger.Lazy
 import okhttp3.Interceptor
 import okhttp3.Request
 import okhttp3.Response
-import org.simple.clinic.facility.FacilityRepository
-import java.util.UUID
+import org.simple.clinic.util.toNullable
 import javax.inject.Inject
 
 class LoggedInUserHttpInterceptor @Inject constructor(
-    private val facilityRepository: FacilityRepository,
-    // TODO(vs): 2019-11-12 Fix this when creating multiple dagger scopes
-    private val userSession: Lazy<UserSession>
+    private val userSession: UserSession
 ) : Interceptor {
 
   override fun intercept(chain: Interceptor.Chain?): Response {
     val originalRequest = chain!!.request()
 
-    val user = userSession.get().loggedInUserImmediate()
-    val (accessToken) = userSession.get().accessToken()
+    val user = userSession.loggedInUserImmediate()
+    val accessToken = userSession.accessToken().toNullable()
 
-    var facilityUuid: UUID? = null
-    if (user != null) {
-      facilityUuid = facilityRepository.currentFacilityUuid()
-    }
-
-    return if (user != null && accessToken.isNullOrBlank().not() && facilityUuid != null) {
-      chain.proceed(addHeaders(originalRequest, accessToken!!, user, facilityUuid))
+    return if (user != null && accessToken.isNullOrBlank().not()) {
+      chain.proceed(addHeaders(originalRequest, accessToken!!, user))
     } else {
       chain.proceed(originalRequest)
     }
   }
 
-  private fun addHeaders(originalRequest: Request, accessToken: String, user: User, facilityUuid: UUID): Request {
+  private fun addHeaders(originalRequest: Request, accessToken: String, user: User): Request {
     return originalRequest.newBuilder()
         .addHeader("Authorization", "Bearer $accessToken")
         .addHeader("X-USER-ID", user.uuid.toString())
-        .addHeader("X-FACILITY-ID", facilityUuid.toString())
+        .addHeader("X-FACILITY-ID", user.currentFacilityUuid.toString())
         .build()
   }